feat: Get ship counts directly from rle file rather than hdf5 database.
Reading from hdf5 database is only slightly faster ~ 1 minute.
And since we only need the raw ship masks once at the beginning of the
project, it doesn't make sense to require the user to generate
what would eventually be a 25GB hdf5 database file just to read
from it once.